Extract query resolving to more generic functions
