+++ extensions/ql-vscode/test/vscode-tests/no-workspace/language-support/contextual/query-resolver.test.ts
@@ -5,23 +5,11 @@ import { getErrorMessage } from "../../../../../src/common/helpers-pure";
 
 import * as log from "../../../../../src/common/logging/notifications";
 import * as workspaceFolders from "../../../../../src/common/vscode/workspace-folders";
-import * as qlpack from "../../../../../src/databases/qlpack";
-import {
-  KeyType,
-  qlpackOfDatabase,
-  resolveQueries,
-} from "../../../../../src/language-support";
+import { KeyType, resolveQueries } from "../../../../../src/language-support";
 import { CodeQLCliServer } from "../../../../../src/codeql-cli/cli";
-import { mockDatabaseItem, mockedObject } from "../../../utils/mocking.helpers";
+import { mockedObject } from "../../../utils/mocking.helpers";
 
 describe("queryResolver", () => {
-  let getQlPackForDbschemeSpy: jest.SpiedFunction<
-    typeof qlpack.getQlPackForDbscheme
-  >;
-  let getPrimaryDbschemeSpy: jest.SpiedFunction<
-    typeof qlpack.getPrimaryDbscheme
-  >;
-
   const resolveQueriesInSuite = jest.fn();
 
   const mockCli = mockedObject<CodeQLCliServer>({
@@ -29,16 +17,6 @@ describe("queryResolver", () => {
   });
 
   beforeEach(() => {
-    getQlPackForDbschemeSpy = jest
-      .spyOn(qlpack, "getQlPackForDbscheme")
-      .mockResolvedValue({
-        dbschemePack: "dbschemePack",
-        dbschemePackIsLibraryPack: false,
-      });
-    getPrimaryDbschemeSpy = jest
-      .spyOn(qlpack, "getPrimaryDbscheme")
-      .mockResolvedValue("primaryDbscheme");
-
     jest
       .spyOn(workspaceFolders, "getOnDiskWorkspaceFolders")
       .mockReturnValue([]);
@@ -68,7 +46,7 @@ describe("queryResolver", () => {
           queries: ".",
           include: {
             kind: "definitions",
-            "tags contain": "ide-contextual-queries/local-definitions",
+            "tags contain": ["ide-contextual-queries/local-definitions"],
           },
         },
       ]);
@@ -87,31 +65,9 @@ describe("queryResolver", () => {
         expect(true).toBe(false);
       } catch (e) {
         expect(getErrorMessage(e)).toBe(
-          'No definitions queries (tagged "ide-contextual-queries/local-definitions") could be found in the current library path (tried searching the following packs: my-qlpack). Try upgrading the CodeQL libraries. If that doesn\'t work, then definitions queries are not yet available for this language.',
+          'No definitions queries (kind "definitions", tagged "ide-contextual-queries/local-definitions") could be found in the current library path (tried searching the following packs: my-qlpack). Try upgrading the CodeQL libraries. If that doesn\'t work, then definitions queries are not yet available for this language.',
         );
       }
-    });
-  });
-
-  describe("qlpackOfDatabase", () => {
-    it("should get the qlpack of a database", async () => {
-      getQlPackForDbschemeSpy.mockResolvedValue({
-        dbschemePack: "my-qlpack",
-        dbschemePackIsLibraryPack: false,
-      });
-      const db = mockDatabaseItem({
-        contents: {
-          datasetUri: {
-            fsPath: "/path/to/database",
-          },
-        },
-      });
-      const result = await qlpackOfDatabase(mockCli, db);
-      expect(result).toEqual({
-        dbschemePack: "my-qlpack",
-        dbschemePackIsLibraryPack: false,
-      });
-      expect(getPrimaryDbschemeSpy).toBeCalledWith("/path/to/database");
     });
   });
 });
